resolve getNested* to cosmology/utils
