+++ packages/ast/src/clients/lcd/aggregated/convert.test.ts
@@ -1,5 +1,6 @@
 import { createAggregatedLCDClient } from '../class';
-import { ProtoStore, traverse, getNestedProto } from '@cosmology/proto-parser';
+import { traverse } from '@cosmology/proto-parser'
+import { getNestedProto } from '@cosmology/utils';
 import {
   defaultTelescopeOptions,
   ProtoRef,
