+++ packages/ast/src/clients/lcd/class/lcd.arrowFunc.test.ts
@@ -1,5 +1,6 @@
 import { createLCDClient } from './lcd';
-import { traverse, getNestedProto } from '@cosmology/proto-parser'
+import { traverse } from '@cosmology/proto-parser'
+import { getNestedProto } from '@cosmology/utils'
 import { ProtoService } from '@cosmology/types';
 import { GenericParseContext } from '../../../encoding';
 import { getTestProtoStore, expectCode } from '../../../../test-utils';
