+++ packages/ast/src/encoding/proto/implements/__tests__/interface-decoder.test.ts
@@ -1,4 +1,4 @@
-import { getNestedProto } from '@cosmology/proto-parser';
+import { getNestedProto } from '@cosmology/utils';
 import { InterfaceTypeUrlMap } from '@cosmology/types';
 import { expectCode, getTestProtoStore } from '../../../../../test-utils/'
 import { ProtoParseContext } from '../../../context';
