+++ packages/ast/src/encoding/proto/interface/__tests__/interface.type-encoded.test.ts
@@ -1,4 +1,4 @@
-import { getNestedProto } from '@cosmology/proto-parser';
+import { getNestedProto } from '@cosmology/utils';
 import { expectCode, getTestProtoStore } from '../../../../../test-utils'
 import { ProtoParseContext } from '../../../context';
 import { createSDKType, createProtoType, createProtoInterfaceEncodedType } from '..';
