+++ packages/ast/src/encoding/proto/proto.enum.spec.ts
@@ -4,7 +4,7 @@ import {
     createProtoEnumToJSON
 } from './enums';
 import struct from '../../../../../__fixtures__/proto-json/google/protobuf/struct.json'
-import { getNestedProto } from '@cosmology/proto-parser';
+import { getNestedProto } from '@cosmology/utils';
 import { ProtoParseContext } from '@cosmology/ast';
 import { expectCode, getTestProtoStore, defaultTelescopeOptions } from '../../../test-utils';
