+++ packages/ast/src/encoding/proto/proto.type.implements.spec.ts
@@ -1,6 +1,6 @@
 import { createProtoType, createCreateProtoType } from './interface';
 
-import { getNestedProto } from '@cosmology/proto-parser'
+import { getNestedProto } from '@cosmology/utils'
 import { ProtoParseContext } from '../context';
 import { getTestProtoStore, expectCode, printCode } from '../../../test-utils';
