+++ packages/parser/__tests__/misc/akash.test.ts
@@ -1,7 +1,7 @@
 import { ProtoRef } from '@cosmology/types';
 import { ProtoStore } from '../../src/index'
 import { traverse } from '../../src';
-import { getNested } from '../../src/utils'
+import { getNested } from '../../src'
 import { getTestProtoStore } from '../../test-utils';
 
 const store = getTestProtoStore();
