+++ packages/parser/types/index.d.ts
@@ -1,5 +1,6 @@
-export * from './store';
-export * from './resolver';
-export * from './utils';
-export * from './traverse';
-export * from './lookup';
+export { getNestedProto, getNestedProtoGeneric, getNested, } from "@cosmology/utils";
+export * from "./store";
+export * from "./resolver";
+export * from "./utils";
+export * from "./traverse";
+export * from "./lookup";
